Update periodTracker.java
with this modification, the program will calculate the average cycle length based on the average of the user-provided cycle length and the length of the previous cycle. then, it will predict the next period and ovulation date based on this average cycle length. this should provide a more accurate prediction, accounting for variations in the length of menstrual cycles.